Add files via upload
The puzzle is made up of four dice with colored dots. Each of the dice has red, green, blue, or yellow arranged in a certain way and no two dice are the same. The objective is to arrange all four dice, stacked side-by-side, so that there will be four different colors appearing on the front, top, back, and bottom. This python program cycles through every possible way we can arrange the four cubes until a solution is reached. In total, there are 82944 ways of stacking these dice and there is only one solution.