Add files via upload
The puzzle is made up of four cubes with colored dots. Each of the cube has red, green, blue, or yellow arranged in a certain way and no two cubes are the same. The objective is to arrange all four cubes, stacked side-by-side, so that there will be four different colors appearing on the front, top, back, and bottom. 

I show a python program that cycles through possible arrangements of the four cubes until a solution is reached. In total, there are 82,944 ways of stacking these cubes and there is only one solution.